% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_3_complete_1_setup_1_domains_fct_metadata.R
\name{create_domain_var_val_df}
\alias{create_domain_var_val_df}
\title{Create data frame of domain var values combinations}
\usage{
create_domain_var_val_df(path, domain_vars)
}
\arguments{
\item{path}{Path. Path to the raw questionnaire metadata data frame.}

\item{domain_vars}{Character vector. Names of domain vars.}
}
\value{
Data frame.
Columns are domain variables.
Contents are options.
Extent is the combination of all answer options for all domain variables.
}
\description{
Create data frame of domain var values combinations
}
