% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_3_complete_1_setup_1_domains_fct_metadata.R
\name{make_vars_options}
\alias{make_vars_options}
\title{Make display variable options}
\usage{
make_vars_options(path, var_types)
}
\arguments{
\item{path}{Character. Path to questionnaire variables file (qnr_vars.rds).}

\item{var_types}{Character. Types of Survey Solutions variable to include.
Use types from SuSo JSON file.}
}
\value{
Character vector. Set of variable descriptions
}
\description{
Compose domain as `{varname} : {variable_description}`
}
